refactor(exception): 디렉토리 구조 정리 및 클래스명 변경
