--- src/main/java/wooteco/subway/exception/InvalidDeleteSectionException.java
@@ -1,7 +1,0 @@
-package wooteco.subway.exception;
-
-public class InvalidDeleteSectionException extends SubwayException {
-    public InvalidDeleteSectionException() {
-        super("삭제할 수 없는 구간입니다.");
-    }
-}
