--- src/main/java/wooteco/subway/exception/NoLineException.java
@@ -1,8 +1,0 @@
-package wooteco.subway.exception;
-
-public class NoLineException extends SubwayException {
-
-    public NoLineException() {
-        super("존재하지 않는 노선입니다.");
-    }
-}
