--- src/main/java/wooteco/subway/exception/NoStationException.java
@@ -1,10 +1,0 @@
-package wooteco.subway.exception;
-
-public class NoStationException extends SubwayException {
-
-    public NoStationException() {
-        super("존재하지 않는 역입니다.");
-    }
-}
-
-
