+++ src/main/java/wooteco/subway/exception/line/LineDuplicationException.java
@@ -1,4 +1,6 @@
-package wooteco.subway.exception;
+package wooteco.subway.exception.line;
+
+import wooteco.subway.exception.SubwayException;
 
 public class LineDuplicationException extends SubwayException {
     public LineDuplicationException() {
