+++ src/main/java/wooteco/subway/exception/line/LineNonexistenceException.java
@@ -1,0 +1,10 @@
+package wooteco.subway.exception.line;
+
+import wooteco.subway.exception.SubwayException;
+
+public class LineNonexistenceException extends SubwayException {
+
+    public LineNonexistenceException() {
+        super("존재하지 않는 노선입니다.");
+    }
+}
