+++ src/main/java/wooteco/subway/exception/section/SectionDeletionException.java
@@ -1,0 +1,9 @@
+package wooteco.subway.exception.section;
+
+import wooteco.subway.exception.SubwayException;
+
+public class SectionDeletionException extends SubwayException {
+    public SectionDeletionException() {
+        super("삭제할 수 없는 구간입니다.");
+    }
+}
