+++ src/main/java/wooteco/subway/exception/section/SectionDuplicationException.java
@@ -1,4 +1,6 @@
-package wooteco.subway.exception;
+package wooteco.subway.exception.section;
+
+import wooteco.subway.exception.SubwayException;
 
 public class SectionDuplicationException extends SubwayException {
     public SectionDuplicationException() {
