+++ src/main/java/wooteco/subway/exception/station/StationDuplicationException.java
@@ -1,4 +1,6 @@
-package wooteco.subway.exception;
+package wooteco.subway.exception.station;
+
+import wooteco.subway.exception.SubwayException;
 
 public class StationDuplicationException extends SubwayException {
     public StationDuplicationException() {
